fix(mouse): rename builtin Mouse library to Mouse_STM32 to avoid conflict Signed-off-by: Aymane Bahssain <aymane.bahssain@st.com>
Rename and update includes for the builtin STM32 mouse library to `Mouse_STM32` to avoid name
conflicts with the official Arduino `Mouse` library
