+++ libraries/CMakeLists.txt
@@ -4,7 +4,7 @@ add_subdirectory(CMSIS_DSP)
 add_subdirectory(EEPROM)
 add_subdirectory(IWatchdog)
 add_subdirectory(Keyboard)
-add_subdirectory(Mouse)
+add_subdirectory(Mouse_STM32)
 add_subdirectory(RGB_LED_TLC59731)
 add_subdirectory(SPI)
 add_subdirectory(Servo)
