+++ libraries/Mouse_STM32/library.properties
@@ -1,4 +1,4 @@
-name=Mouse
+name=Mouse_STM32
 version=1.1.0
 author=Arduino, stm32duino
 maintainer=stm32duino
